+++ packages/analytics/src/expand.js
@@ -2,7 +2,6 @@ import { resolve  as resolvePath } from 'path';
 import { tmpdir } from 'os';
 import get from 'lodash.get';
 import set from 'lodash.set';
-import { createStore } from '@ezs/store';
 import cacache from 'cacache';
 import each from 'async-each-series';
 import makeDir from 'make-dir';
@@ -33,25 +32,24 @@ async function mergeWith(data, feed) {
     const {
         store,
         cachePath,
-        stack,
-        bufferID,
     } = this.getEnv();
     const { id, value } = data;
     const path = this.getParam('path');
     try {
-        const obj = await store.cut(id);
-        if (obj === null) {
+        const obj = store[id];
+        delete store[id];
+        if (obj === undefined || obj === null) {
             throw new Error('id was corrupted');
         }
         const source = get(obj, path);
         if (cachePath && source) {
             await cachePut(cachePath, source, value);
         }
-        delete stack[bufferID][id];
         set(obj, path, value);
         return feed.send(obj);
     } catch (e) {
-        return feed.stop(e);
+        // avoid to break the pipe
+        return feed.send(e);
     }
 }
 
@@ -59,21 +57,19 @@ async function drainWith(data, feed) {
     if (this.isLast()) {
         const {
             store,
-            stack,
-            bufferID,
         } = this.getEnv();
-
         return each(
-            Object.keys(stack[bufferID]),
+            Object.keys(store),
             async (cur, next) => {
                 let obj;
                 try {
-                    obj = await store.get(cur);
+                    obj = store[cur];
+                    delete store[cur];
                 } catch (e) {
                     feed.write(e);
                 }
-                if (obj === null) {
-                    feed.stop(new Error(`Unable to find ${cur} in the store ${store.id()}`));
+                if (obj === undefined || obj === null) {
+                    feed.stop(new Error(`Unable to find ${cur} in the store`));
                 } else {
                     feed.write(obj);
                 }
@@ -138,9 +134,7 @@ export default async function expand(data, feed) {
         const cacheName = this.getParam('cacheName');
 
         if (!this.store) {
-            const location = this.getParam('location');
-            this.store = createStore(ezs, 'expand', location);
-            await this.store.reset();
+            this.store = {}
             this.flows = [];
         }
         if (!this.createStatements) {
@@ -163,48 +157,39 @@ export default async function expand(data, feed) {
         }
 
         if (!this.buffer2stream) {
-            this.buffer2stream = (bufferID) => {
+            this.buffer2stream = () => {
                 const statements = this.createStatements();
                 const stream = ezs.createStream(ezs.objectMode());
                 const output = ezs.createPipeline(stream, statements)
                     .pipe(ezs(mergeWith, { path }, {
                         store: this.store,
                         cachePath: this.cachePath,
-                        stack: this.stack,
-                        bufferID,
                     }))
                     .pipe(ezs(drainWith, { path }, {
                         store: this.store,
                         cachePath: this.cachePath,
-                        stack: this.stack,
-                        bufferID,
-                    }));
+                    }))
+                ;
                 const input = Array.from(this.buffer);
                 this.buffer = [];
-                this.bufferID += 1;
-                this.stack[this.bufferID] = {};
 
                 each(input, (cur, next) => ezs.writeTo(stream, cur, next), () => stream.end());
                 return output;
             };
         }
         if (!this.buffer) {
-            this.stack = [];
             this.buffer = [];
-            this.bufferID = 0;
-            this.stack[this.bufferID] = {};
         }
 
         // Processing
 
         if (this.isLast()) {
             if (this.buffer && this.buffer.length > 0) {
-                const strm = this.buffer2stream(this.bufferID);
+                const strm = this.buffer2stream();
                 this.flows.push(feed.flow(strm));
             }
             await Promise.all(this.flows);
-            await this.store.close();
-
+            delete this.store;
             return feed.close();
         }
         const value = get(data, path);
@@ -221,18 +206,17 @@ export default async function expand(data, feed) {
         const id = this.getIndex().toString().padStart(20, '0');
         const size = Number(this.getParam('size', 1));
 
-        await this.store.put(id, data);
+        this.store[id] = data;
 
-        this.stack[this.bufferID][id] = true;
         this.buffer.push(core(id, value));
         if (this.buffer.length >= size) {
-            const strm = this.buffer2stream(this.bufferID);
+            const strm = this.buffer2stream();
             return this.flows.push(feed.flow(strm));
         }
         return feed.end();
     } catch (e) {
         if (this.store) {
-            this.store.close();
+            delete this.store;
         }
         return feed.stop(e);
     }