+++ packages/core/src/statements/index.js
@@ -29,6 +29,7 @@ import swing from './swing';
 import loop from './loop';
 import map from './map';
 import identify from './identify';
+import throttle from './throttle';
 
 export default {
     extract,
@@ -62,4 +63,5 @@ export default {
     loop,
     map,
     identify,
+    throttle,
 };